osband thompson sampling notebook
To run the notebook:

Make sure you're in the ts_tutorial-master folder.

Open the notebook and click the "Open in Colab" button.

Should be straightforward to run each of the modules from there. Run this notebook and NOT the one in the Ian Osband repo since that one had a few commands that no longer work. The one in this repo has been updated and works.